+++ sentry/src/test/java/io/sentry/SentryExceptionFactoryTest.kt
@@ -1,28 +1,65 @@
 package io.sentry
 
+import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.exception.ExceptionMechanismException
 import io.sentry.protocol.Mechanism
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertNull
 import kotlin.test.assertTrue
 
 class SentryExceptionFactoryTest {
-    private val sut = SentryExceptionFactory(SentryStackTraceFactory(listOf("io.sentry"), listOf()))
+    private class Fixture {
+
+        fun getSut(stackTraceFactory: SentryStackTraceFactory = SentryStackTraceFactory(listOf("io.sentry"), listOf())): SentryExceptionFactory {
+            return SentryExceptionFactory(stackTraceFactory)
+        }
+    }
+
+    private val fixture = Fixture()
 
     @Test
     fun `when getSentryExceptions is called passing an Exception, not empty result`() {
         val exception = Exception("Exception")
-        assertTrue(sut.getSentryExceptions(exception).size > 0)
+        assertTrue(fixture.getSut().getSentryExceptions(exception).isNotEmpty())
     }
 
     @Test
     fun `when getSentryExceptions is called passing an Exception, it should set its fields`() {
         val exception = Exception("Exception")
-        val sentryExceptions = sut.getSentryExceptions(exception)
+        val sentryExceptions = fixture.getSut().getSentryExceptions(exception)
         assertEquals("Exception", sentryExceptions[0].type)
         assertEquals("Exception", sentryExceptions[0].value)
         assertEquals("java.lang", sentryExceptions[0].module)
-        assertTrue(sentryExceptions[0].stacktrace.frames.size > 0)
+        assertTrue(sentryExceptions[0].stacktrace.frames.isNotEmpty())
+    }
+
+    @Test
+    fun `when frames are null, do not set a stack trace object`() {
+        val stackTraceFactory = mock<SentryStackTraceFactory>()
+        whenever(stackTraceFactory.getStackFrames(any())).thenReturn(null)
+
+        val sut = fixture.getSut(stackTraceFactory)
+        val exception = Exception("Exception")
+
+        val sentryExceptions = sut.getSentryExceptions(exception)
+
+        assertNull(sentryExceptions[0].stacktrace)
+    }
+
+    @Test
+    fun `when frames are empty, do not set a stack trace object`() {
+        val stackTraceFactory = mock<SentryStackTraceFactory>()
+        whenever(stackTraceFactory.getStackFrames(any())).thenReturn(emptyList())
+
+        val sut = fixture.getSut(stackTraceFactory)
+        val exception = Exception("Exception")
+
+        val sentryExceptions = sut.getSentryExceptions(exception)
+
+        assertNull(sentryExceptions[0].stacktrace)
     }
 
     @Test
@@ -35,7 +72,7 @@ class SentryExceptionFactoryTest {
 
         val throwable = ExceptionMechanismException(mechanism, error, null)
 
-        val sentryExceptions = sut.getSentryExceptions(throwable)
+        val sentryExceptions = fixture.getSut().getSentryExceptions(throwable)
         assertEquals("anr", sentryExceptions[0].mechanism.type)
         assertEquals(false, sentryExceptions[0].mechanism.isHandled)
     }
@@ -43,7 +80,7 @@ class SentryExceptionFactoryTest {
     @Test
     fun `when exception is nested, it should be sorted oldest to newest`() {
         val exception = Exception("message", Exception("cause"))
-        val queue = sut.extractExceptionQueue(exception)
+        val queue = fixture.getSut().extractExceptionQueue(exception)
 
         assertEquals("cause", queue.first.value)
         assertEquals("message", queue.last.value)
@@ -52,13 +89,13 @@ class SentryExceptionFactoryTest {
     @Test
     fun `when getSentryExceptions is called passing an Inner exception, not empty result`() {
         val exception = InnerClassThrowable(InnerClassThrowable())
-        val queue = sut.extractExceptionQueue(exception)
+        val queue = fixture.getSut().extractExceptionQueue(exception)
         assertEquals("SentryExceptionFactoryTest\$InnerClassThrowable", queue.first.type)
     }
 
     @Test
     fun `when getSentryExceptions is called passing an anonymous exception, not empty result`() {
-        val queue = sut.extractExceptionQueue(anonymousException)
+        val queue = fixture.getSut().extractExceptionQueue(anonymousException)
         assertEquals("SentryExceptionFactoryTest\$anonymousException\$1", queue.first.type)
     }
 
@@ -66,7 +103,7 @@ class SentryExceptionFactoryTest {
     fun `when exception has no mechanism, it should get and set the current threadId`() {
         val threadId = Thread.currentThread().id
         val exception = Exception("message", Exception("cause"))
-        val queue = sut.extractExceptionQueue(exception)
+        val queue = fixture.getSut().extractExceptionQueue(exception)
 
         assertEquals(threadId, queue.first.threadId)
     }
@@ -79,7 +116,7 @@ class SentryExceptionFactoryTest {
         val thread = Thread()
         val throwable = ExceptionMechanismException(mechanism, exception, thread)
 
-        val queue = sut.extractExceptionQueue(throwable)
+        val queue = fixture.getSut().extractExceptionQueue(throwable)
 
         assertEquals(thread.id, queue.first.threadId)
     }